app-bot/requirements.txt
aiofiles
aiohttp
aiohttp_session
asyncpg
autopep8
discord.py
discord.ext.oauth
ipython
jishaku
matplotlib
matplotlib_venn
nasapy
numpy
openpyxl
pandas
parse
Pillow
plotly
psycopg2-binary
Pydap
pydub
pyparsing
python-dotenv
python_dateutil
requests
rich
schedule
wolframalpha